gcloud/cloudfunctions/cloudFunction1/requirements.txt
# Function dependencies, for example:
# package>=version
requests
google-cloud-pubsub